# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "易宿酒店预订平台 API",
    "version": "1.0.0",
    "description": "基于 Node.js Express 的酒店预订系统后端接口文档"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": ["username", "password", "role"],
        "properties": {
          "username": { "type": "string", "example": "merchant_01" },
          "password": { "type": "string", "example": "password123", "description": "加密存储" },
          "role": {
            "type": "string",
            "enum": ["user", "merchant", "admin"],
            "default": "user",
            "example": "merchant"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "unknown"],
            "default": "unknown",
            "example": "male"
          },
          "avatar": { "type": "string", "example": "/uploads/avatars/default.png" },
          "bio": { "type": "string", "example": "热爱旅行" }
        }
      },
...
      "RoomType": {
        "type": "object",
        "required": ["hotelId", "title", "price"],
        "properties": {
          "hotelId": { "type": "string", "description": "关联酒店ID", "example": "609c1..." },
          "title": { "type": "string", "example": "云端大床房" },
          "price": { "type": "number", "description": "实际价格", "example": 5000 },
          "originalPrice": { "type": "number", "description": "原价(可选)", "example": 8888 },
          "capacity": { "type": "integer", "default": 2, "example": 2 },
          "bedInfo": { "type": "string", "example": "1张2米特大床" },
          "size": { "type": "string", "example": "60m²" },
          "stock": { "type": "integer", "description": "库存 (必填)", "example": 10 },
          "images": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["/uploads/room1.jpg"]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/users/profile": {
      "get": {
        "summary": "获取个人详细资料",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/User" } }
            }
          }
        }
      },
      "put": {
        "summary": "修改个人资料 (头像/性别/简介)",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": { "type": "string", "example": "/uploads/new-avatar.jpg" },
                  "gender": { "type": "string", "enum": ["male", "female", "unknown"] },
                  "bio": { "type": "string", "example": "这个人很懒，什么都没写" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "修改成功" }
        }
      }
    },
    "/api/reviews": {
      "post": {
        "summary": "发布评价",
        "tags": ["Reviews"],
        "security": [{ "bearerAuth": [] }],
        "description": "评价发布后，会自动重新计算并更新酒店的总评分",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hotelId", "rating", "content"],
                "properties": {
                  "hotelId": { "type": "string" },
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5, "description": "1-5分" },
                  "content": { "type": "string", "example": "房间很干净，服务很好！" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "评价成功" },
          "400": { "description": "已评价过" }
        }
      }
    },
    "/api/reviews/{hotelId}": {
      "get": {
        "summary": "获取某酒店的所有评价",
        "tags": ["Reviews"],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "userId": { "type": "object", "properties": { "username": { "type": "string" }, "avatar": { "type": "string" } } },
                      "rating": { "type": "integer" },
                      "content": { "type": "string" },
                      "createdAt": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rooms": {
      "post": {
        "summary": "添加房型 (仅商户)",
        "tags": ["Rooms"],
        "description": "添加后会自动更新酒店的最低价(minPrice)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RoomType" } }
          }
        },
        "responses": {
          "200": { "description": "添加成功" },
          "401": { "description": "无权操作" }
        }
      }
    },
    "/api/rooms/{hotelId}": {
      "get": {
        "summary": "获取某酒店的所有房型",
        "tags": ["Rooms"],
        "description": "公共接口，用于详情页展示，按价格从低到高排序",
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/RoomType" } }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{id}": {
      "delete": {
        "summary": "删除房型 (仅商户)",
        "tags": ["Rooms"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "删除成功" }
        }
      }
    },
...
  }
}